Refactor: 레이아웃 관련 컴포넌트 파일 위치 변경
