+++ src/pages/MainPage.jsx
@@ -1,8 +1,8 @@
 
-import Header from "../components/Header";
+import Header from "../components/Layout/Common/Header.jsx";
 import Main from "../components/Main/Main";
-import Footer from "../components/Footer";
-import Banner from "../components/Banner";
+import Footer from "../components/Layout/Footer.jsx";
+import Banner from "../components/Layout/Banner.jsx";
 
 function MainPage() {
   return (
